refactor: extract valid MBR greater GBR part to an additional func
