+++ frontend/src/pages/ProfileCreate/index.tsx
@@ -11,7 +11,7 @@ import ProfileFormBasic from "./ProfileFormBasic";
 import ProfileFormUeAmbr from "./ProfileFormUeAmbr";
 import ProfileFormSessions from "./ProfileFormSessions";
 import { ProfileMapperImpl, FlowsMapperImpl } from "../../lib/dtos/profile";
-import { parseDataRate } from "../../lib/utils";
+import { validateMBRGreaterThanGBR } from "../../lib/utils";
 
 function FormHOC(Component: React.ComponentType<any>) {
     return function (props: any) {
@@ -68,22 +68,10 @@ function ProfileCreate() {
     const profileMapper = new ProfileMapperImpl(new FlowsMapperImpl());
     const profile = profileMapper.mapFromDto(data);
 
-    for (let i = 0; i < profile.QosFlows.length; i++) {
-      const qosFlow = profile.QosFlows[i];
-      const gbrDL = parseDataRate(qosFlow.gbrDL);
-      const mbrDL = parseDataRate(qosFlow.mbrDL);
-      const gbrUL = parseDataRate(qosFlow.gbrUL);
-      const mbrUL = parseDataRate(qosFlow.mbrUL);
-
-      if (gbrDL && mbrDL && gbrDL >= mbrDL) {
-          alert("In S-NSSAI " + qosFlow.snssai + "'s Flow Rule " + (i+1) + "\nDownlink MBR must be greater than Downlink GBR");
-          return;
-      }
-
-      if (gbrUL && mbrUL && gbrUL >= mbrUL) {
-          alert("In S-NSSAI " + qosFlow.snssai + "'s Flow Rule " + (i+1) + "\nUplink MBR must be greater than Uplink GBR");
-          return;
-      }
+    const validation = validateMBRGreaterThanGBR(profile.QosFlows);
+    if (!validation.isValid) {
+      alert(validation.error);
+      return;
     }
 
     axios
@@ -116,22 +104,10 @@ function ProfileCreate() {
     const profileMapper = new ProfileMapperImpl(new FlowsMapperImpl());
     const profile = profileMapper.mapFromDto(data);
 
-    for (let i = 0; i < profile.QosFlows.length; i++) {
-      const qosFlow = profile.QosFlows[i];
-      const gbrDL = parseDataRate(qosFlow.gbrDL);
-      const mbrDL = parseDataRate(qosFlow.mbrDL);
-      const gbrUL = parseDataRate(qosFlow.gbrUL);
-      const mbrUL = parseDataRate(qosFlow.mbrUL);
-
-      if (gbrDL && mbrDL && gbrDL >= mbrDL) {
-          alert("In S-NSSAI " + qosFlow.snssai + "'s Flow Rule " + (i+1) + "\nDownlink MBR must be greater than Downlink GBR");
-          return;
-      }
-
-      if (gbrUL && mbrUL && gbrUL >= mbrUL) {
-          alert("In S-NSSAI " + qosFlow.snssai + "'s Flow Rule " + (i+1) + "\nUplink MBR must be greater than Uplink GBR");
-          return;
-      }
+    const validation = validateMBRGreaterThanGBR(profile.QosFlows);
+    if (!validation.isValid) {
+      alert(validation.error);
+      return;
     }
 
     axios
